Moved deleteAnnotationModal to a separate component
