+++ src/components/viewer/Viewer.tsx
@@ -3,33 +3,27 @@ import classNames from 'classnames';
 import { connect } from 'react-redux';
 import { createResource, deleteResource } from 'redux-json-api';
 import Widget from 'components/widget';
-import { Button, Modal } from 'semantic-ui-react';
 
 import ViewerItem from './ViewerItem';
 import styles from './Viewer.scss';
 import { selectViewerState } from 'store/widgets/selectors';
-import { hideViewer, hideViewerDeleteModal, openViewerDeleteModal, showEditorAnnotation } from 'store/widgets/actions';
+import { hideViewer, openViewerDeleteModal, showEditorAnnotation } from 'store/widgets/actions';
 import { AnnotationAPIModel } from 'api/annotations';
 import { PPScopeClass, PPViewerIndirectChildClass } from 'class_consts.ts';
-import Timer = NodeJS.Timer;
-import Highlighter from 'core/Highlighter';
-import { mouseOverViewer } from '../../store/widgets/actions';
+import { mouseOverViewer } from 'store/widgets/actions';
+import DeleteAnnotationModal from './DeleteAnnotationModal';
 
 interface IViewerProps {
   locationX: number;
   locationY: number;
   annotations: AnnotationAPIModel[];
-  deleteModalId: string;
+
   deleteModalOpen: boolean;
 
-  onModalClose: (Event) => void;
-
-  deleteAnnotation: (instance: AnnotationAPIModel) => Promise<object>;
   showEditorAnnotation: (x: number, y: number, id?: string) => void;
   hideViewer: () => void;
   mouseOverViewer: (value: boolean) => void;
   openViewerDeleteModal: (id: string) => void;
-  hideViewerDeleteModal: () => void;
 }
 
 @connect(
@@ -39,7 +33,6 @@ interface IViewerProps {
       locationX,
       locationY,
       deleteModal: {
-        deleteModalId,
         deleteModalOpen,
       },
       annotations,
@@ -49,7 +42,6 @@ interface IViewerProps {
       visible,
       locationX,
       locationY,
-      deleteModalId,
       deleteModalOpen,
       annotations,
     };
@@ -59,8 +51,6 @@ interface IViewerProps {
     hideViewer,
     mouseOverViewer,
     openViewerDeleteModal,
-    hideViewerDeleteModal,
-    deleteAnnotation: deleteResource,
   },
 )
 export default class Viewer extends React.Component<Partial<IViewerProps>, {}> {
@@ -87,19 +77,6 @@ export default class Viewer extends React.Component<Partial<IViewerProps>, {}> {
 
   onItemDelete = (id: string) => {
     this.props.openViewerDeleteModal(id);
-  }
-
-  onItemConfirmedDelete = (e) => {
-    const annotation = this.props.annotations.find(a => a.id === this.props.deleteModalId);
-    this.props.deleteAnnotation(annotation)
-      .catch((errors) => {
-        console.log(errors);
-      });
-  }
-
-  setDeleteModalClosed = (e) => {
-    this.props.hideViewerDeleteModal();
-    this.props.mouseOverViewer(false);
   }
 
   handleMouseLeave = (e) => {
@@ -133,28 +110,6 @@ export default class Viewer extends React.Component<Partial<IViewerProps>, {}> {
     });
   }
 
-  renderDeleteModal() {
-    return (
-      <Modal
-        size="mini"
-        className={PPScopeClass}
-        open={this.props.deleteModalOpen}
-      >
-        <Modal.Content>
-          <p>Czy na pewno chcesz usunąć przypis?</p>
-        </Modal.Content>
-        <Modal.Actions>
-          <Button onClick={this.setDeleteModalClosed} size="tiny" negative={true}>
-            Nie
-          </Button>
-          <Button onClick={this.onItemConfirmedDelete} size="tiny" positive={true}>
-            Tak
-          </Button>
-        </Modal.Actions>
-      </Modal>
-    );
-  }
-
   render() {
     return (
       <Widget
@@ -169,7 +124,7 @@ export default class Viewer extends React.Component<Partial<IViewerProps>, {}> {
         <ul className={styles.annotationItems}>
           {this.renderItems()}
         </ul>
-        {this.renderDeleteModal()}
+        <DeleteAnnotationModal/>
       </Widget>
     );
   }
